refactor: rename Mena components to standard names
