+++ design_system/src/commonMain/kotlin/net/thechance/mena/designsystem/presentation/component/bottomNavigation/BottomNavigationBarItem.kt
@@ -15,8 +15,8 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.painter.Painter
 import androidx.compose.ui.unit.dp
-import net.thechance.mena.designsystem.presentation.component.icon.MenaIcon
-import net.thechance.mena.designsystem.presentation.component.text.MenaText
+import net.thechance.mena.designsystem.presentation.component.icon.Icon
+import net.thechance.mena.designsystem.presentation.component.text.Text
 import net.thechance.mena.designsystem.presentation.theme.theme.Theme
 
 @Composable
@@ -54,14 +54,14 @@ fun BottomNavigationBarItem(
                     alignment = Alignment.TopCenter
                 )
         ) {
-            MenaIcon(
+            Icon(
                 painter = painter,
                 modifier = Modifier.size(24.dp),
                 contentDescription = title,
             )
 
             if (isSelected) {
-                MenaText(
+                Text(
                     text = title,
                     style = Theme.typography.label.medium,
                     modifier = Modifier.padding(top = 4.dp)
