+++ design_system/src/commonMain/kotlin/net/thechance/mena/designsystem/presentation/component/image/Image.kt
@@ -7,7 +7,7 @@ import androidx.compose.ui.graphics.painter.Painter
 import androidx.compose.ui.layout.ContentScale
 
 @Composable
-fun MenaImage(
+fun Image(
     painter: Painter,
     contentDescription: String?,
     modifier: Modifier = Modifier,
