+++ design_system/src/commonMain/kotlin/net/thechance/mena/designsystem/presentation/component/scaffold/Scaffold.kt
@@ -13,7 +13,7 @@ import androidx.compose.ui.draw.blur
 import androidx.compose.ui.unit.dp
 
 @Composable
-fun MenaScaffold(
+fun Scaffold(
     modifier: Modifier = Modifier,
     topBar: @Composable () -> Unit = {},
     bottomBar: @Composable () -> Unit = {},
