+++ design_system/src/commonMain/kotlin/net/thechance/mena/designsystem/presentation/component/segment/SegmentButton.kt
@@ -20,7 +20,7 @@ import androidx.compose.ui.draw.clip
 import androidx.compose.ui.draw.shadow
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
-import net.thechance.mena.designsystem.presentation.component.text.MenaText
+import net.thechance.mena.designsystem.presentation.component.text.Text
 import net.thechance.mena.designsystem.presentation.theme.theme.Theme
 import org.jetbrains.compose.ui.tooling.preview.Preview
 
@@ -61,7 +61,7 @@ internal fun RowScope.SegmentButton(
             },
         contentAlignment = Alignment.Center,
     ) {
-        MenaText(
+        Text(
             text = option,
             style = Theme.typography.label.medium,
             color = if (isSelected) Theme.colorScheme.shadePrimary else Theme.colorScheme.shadeSecondary,
